Add back direct activation with left click
Right-click on the menu item to show the contextual menu.

The feature was removed upon request by App Store reviewer since it goes
against macOS HIG. Since we have no plans to distribute the application
via MAS (category is overcrowded and the app was rejected several times
because of this) we can re-enable the feature.